requirements.txt
# Core essentials
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Model persistence
joblib>=1.3.0

# Basic visualizations
matplotlib>=3.7.0
seaborn>=0.12.0

# Environment management
python-dotenv>=1.0.0 